+++ sdkjs-plugins/content/mendeley/src/app/csl/abbreviations/index.js
@@ -1,0 +1,3 @@
+// @ts-check
+
+export { AbbreviationsManager } from "./abbreviations-manager.js";