+++ sdkjs-plugins/content/mendeley/src/app/csl/citation/citation-item-data.js
@@ -1385,7 +1385,10 @@ CitationItemData.prototype.setCustom = function (custom) {
     return this;
 };
 
-CitationItemData.prototype.toJSON = function () {
+/**
+ * @param {boolean} [skipUrlForPaperArticles]
+ */
+CitationItemData.prototype.toJSON = function (skipUrlForPaperArticles) {
     var result = {};
     result.id = this._id;
 
@@ -1579,6 +1582,16 @@ CitationItemData.prototype.toJSON = function () {
     if (Object.keys(this._custom).length !== 0) result.custom = this._custom;
     if (this._license !== undefined && this._license !== "")
         result.license = this._license;
+
+    if (
+        skipUrlForPaperArticles &&
+        result.page &&
+        (result.type === "article-journal" ||
+            result.type === "article-magazine" ||
+            result.type === "article-newspaper")
+    ) {
+        delete result.URL;
+    }
 
     return result;
 };