+++ sdkjs-plugins/content/mendeley/src/app/csl/citation/citation-item.js
@@ -245,13 +245,16 @@ CitationItem.prototype.addUri = function (uri) {
     return this;
 };
 
-CitationItem.prototype.toJSON = function () {
+/**
+ * @param {boolean} [skipUrlForPaperArticles]
+ */
+CitationItem.prototype.toJSON = function (skipUrlForPaperArticles) {
     var result = {};
     result.id = this.id;
 
     if (this._itemData) {
         result.itemData = this._itemData.toJSON
-            ? this._itemData.toJSON()
+            ? this._itemData.toJSON(skipUrlForPaperArticles)
             : this._itemData;
     }
     if (this._prefix !== undefined) result.prefix = this._prefix;
@@ -269,9 +272,10 @@ CitationItem.prototype.toJSON = function () {
 
 /**
  * @param {number} index
+ * @param {boolean} [skipUrlForPaperArticles]
  * @returns
  */
-CitationItem.prototype.toFlatJSON = function (index) {
+CitationItem.prototype.toFlatJSON = function (index, skipUrlForPaperArticles) {
     /** @type {OldCitationItem} */
     var oldItem = {
         id: this.id,
@@ -281,7 +285,7 @@ CitationItem.prototype.toFlatJSON = function (index) {
         oldItem["suppress-author"] = this._suppressAuthor;
     }
 
-    let itemDataObject = this._itemData.toJSON();
+    let itemDataObject = this._itemData.toJSON(skipUrlForPaperArticles);
     Object.assign(oldItem, itemDataObject);
 
     if (